Unify transaction data returned by api
